myDressRoom/inference/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
Pillow>=10.0.0
